medication-assistant-backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openai==1.52.2
httpx==0.27.2
APScheduler==3.10.4
python-dateutil==2.9.0.post0
email-validator==2.2.0

